api/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
nltk==3.8.1
sumy==0.11.0
dnspython==2.4.2
python-dotenv==1.0.0
numpy<2.0.0
